backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
anthropic==0.18.1
httpx==0.26.0
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
chromadb==0.4.22
